add new tool: ruaumoko-ascii-map
This is a silly little tool which can be used to visually verify that
the dataset which has been downloaded is sane and being parsed
correctly. When run, it generates a little ASCII-art elevation map. If
the output looks like the world map then you can have some confidence
that the dataset has downloaded correctly.