api: use flask-script for development server
The "idiomatic" way to have a ``ruaumoko-api``-like tool with flask is
to use Flask-Script's manager object. For one line of code, we get the
same support which we previously had along with the ability to specify
bind address and port which is occasionally useful. It also supports
live-reloading of the Python module which is almost always useful for
development.